fix(avo-2410): convert rich text editor state to html before save
https://meemoo.atlassian.net/browse/AVO-2410
