+++ ui/src/react-admin/modules/content-page/views/ContentPageEdit.tsx
@@ -15,6 +15,7 @@ import {
 	ContentPageEditState,
 } from '~modules/content-page/helpers/content-edit.reducer';
 import { useContentTypes } from '~modules/content-page/hooks/useContentTypes';
+import { convertRichTextEditorStatesToHtml } from '~modules/content-page/services/content-page.converters';
 import { ContentPageService } from '~modules/content-page/services/content-page.service';
 import {
 	ContentBlockComponentState,
@@ -246,7 +247,7 @@ const ContentPageEdit: FC<ContentPageEditProps> = ({ id, className }) => {
 			// and we don't want those states to end up in the database
 			const blockConfigs: ContentBlockConfig[] = contentPageState.currentContentPageInfo
 				.content_blocks
-				? ContentPageService.convertRichTextEditorStatesToHtml(
+				? convertRichTextEditorStatesToHtml(
 						contentPageState.currentContentPageInfo.content_blocks
 				  )
 				: [];
